Backend/requirements.txt
# Fitness Management System - Backend Requirements
# Python 3.8+

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
pymysql==1.1.0
sqlalchemy==2.0.23
cryptography==41.0.7

# Authentication
bcrypt==4.1.1

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic[email]==2.5.0

# CORS
python-multipart==0.0.6
